requirements.txt
flask
gunicorn
langchain
langchain-openai
langchain-chroma
python-dotenv
tiktoken
openai
langgraph
python-docx